+++ test/expect/TestScript.test_constant_pooling.expect
@@ -1,0 +1,31 @@
+graph(%x.1 : Dynamic) {
+  %c.1 : int = prim::Constant[value=0]()
+  %2 : int[] = prim::Constant[value=[1]]()
+  %3 : int[] = prim::Constant[value=[0, -1]]()
+  %4 : int = prim::Constant[value=6]()
+  %5 : int[] = prim::Constant[value=[0]]()
+  %6 : int = prim::Constant[value=3]()
+  %e : string = prim::Constant[value="bcd"]()
+  %d : string = prim::Constant[value="abc"]()
+  %a : int = prim::Constant[value=1]()
+  %10 : int = prim::TensorToNum(%x.1)
+  %c : int, %x : Dynamic = prim::If(%10)
+    block0() {
+      -> (%6, %x.1)
+    }
+    block1() {
+      %x.2 : Dynamic = aten::rand(%5, %4, %c.1, %3)
+      %14 : int = prim::TensorToNum(%x.2)
+      %x.4 : Dynamic = prim::If(%14)
+        block0() {
+          %x.3 : Dynamic = aten::rand(%2, %4, %c.1, %3)
+          -> (%x.3)
+        }
+        block1() {
+          -> (%x.2)
+        }
+       = prim::Print(%d, %e, %d)
+      -> (%c.1, %x.4)
+    }
+  return (%a, %6, %c, %x);
+}
